requirements.txt
sseclient-py==1.7.2